+++ src/python/py_imagebufalgo.cpp
@@ -1430,6 +1430,27 @@ IBA_resample_ret (const ImageBuf &src, bool interpolate,
 
 
 bool
+IBA_fit (ImageBuf &dst, const ImageBuf &src,
+         const std::string &filtername = "", float filterwidth = 0.0f,
+         bool exact=false, ROI roi=ROI::All(), int nthreads=0)
+{
+    py::gil_scoped_release gil;
+    return ImageBufAlgo::fit (dst, src, filtername, filterwidth,
+                              exact, roi, nthreads);
+}
+
+ImageBuf
+IBA_fit_ret (const ImageBuf &src,
+             const std::string &filtername = "", float filterwidth = 0.0f,
+             bool exact=false, ROI roi=ROI::All(), int nthreads=0)
+{
+    py::gil_scoped_release gil;
+    return ImageBufAlgo::fit (src, filtername, filterwidth, exact, roi, nthreads);
+}
+
+
+
+bool
 IBA_make_kernel (ImageBuf &dst, const std::string &name,
                  float width, float height, float depth, bool normalize)
 {
@@ -2600,6 +2621,13 @@ void declare_imagebufalgo (py::module &m)
         .def_static("resample", &IBA_resample_ret,
             "src"_a, "interpolate"_a=true,
             "roi"_a=ROI::All(), "nthreads"_a=0)
+
+        .def_static("fit", &IBA_fit,
+            "dst"_a, "src"_a, "filtername"_a="", "filterwidth"_a=0.0f,
+             "exact"_a=false, "roi"_a=ROI::All(), "nthreads"_a=0)
+        .def_static("fit", &IBA_fit_ret,
+            "src"_a, "filtername"_a="", "filterwidth"_a=0.0f,
+            "exact"_a=false, "roi"_a=ROI::All(), "nthreads"_a=0)
 
         .def_static("make_kernel", &IBA_make_kernel,
             "dst"_a, "name"_a, "width"_a, "height"_a,